+++ src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Global } from '@emotion/react';
 import resetStyles from '@styles/resetStyles';
-import GoodsItem from '@components/GoodsItem/GoodItem';
-import Container from './components/Layout/Container';
+import GoodsItem from '@/components/common/GoodsItem/GoodItem';
+import Container from './components/common/Layout/Container';
 
 function App() {
   return (
