+++ src/components/common/GoodsItem/Ranking.tsx
